Separate finite element shape source
